+++ app/src/main/java/com/episode6/headachetracker/ui/navigation/Routes.kt
@@ -14,9 +14,6 @@ sealed interface Route {
     data class FullYear(val year: Int) : Route
 
     @Serializable
-    data object NotesSummary : Route
-
-    @Serializable
     data object Settings : Route
 
     @Serializable